Redo source header menus
