+++ mcweb/frontend/src/features/sources/SourceShow.jsx
@@ -120,26 +120,23 @@ export default function SourceShow() {
         </div>
         )}
         {source.notes && (
-        <p>
+        <div>
           {/* eslint-disable-next-line react/jsx-one-expression-per-line */}
           <b>Notes:</b> {source.notes && renderNotes(source.notes, false)}
-        </p>
+        </div>
         )}
         {(source.alternative_domains[0]) && (
-        <div>
           <div className="row">
             {/* eslint-disable-next-line react/jsx-one-expression-per-line */}
-            <p><b>Alternative Domains</b>:
-              <ul>
-                {source.alternative_domains.map((aD) => (
-                  <li>
-                    {aD.domain}
-                  </li>
-                ))}
-              </ul>
-            </p>
+            <b>Alternative Domains:</b>
+            <ul>
+              {source.alternative_domains.map((aD) => (
+                <li key={aD.domain}>
+                  {aD.domain}
+                </li>
+              ))}
+            </ul>
           </div>
-        </div>
         )}
         <p>
           {/* eslint-disable-next-line react/jsx-one-expression-per-line */}
